chore(frontend): satisfy transcript source budgets
Problem:
CI repolint rejected the running-turn stability change because several already-baselined transcript owners and one race test exceeded their source-size ceilings.

Root cause:
The live-turn split, footer stability hook, pre-paint tail repair, and JSDOM frame harness were implemented inline in files already at their ownership limits.

Fix:
Extract live-turn geometry, stability, tail-settle, and recovery-race setup into focused owners without changing behavior or raising repolint baselines. Keep the measured startup increase behind the smallest 0.1 KiB bundle ratchets.

Verification:
- go run ./tools/repolint
- pnpm test:transcript
- pnpm build
- initial JavaScript gzip: 448.321 KiB / 448.4 KiB
- initial raw JavaScript and CSS: 2415.572 KiB / 2415.7 KiB
