+++ desktop/frontend/src/__tests__/helpers/transcriptRecoveryRaceDom.ts
@@ -1,0 +1,36 @@
+import { JSDOM } from "jsdom";
+import { act } from "react";
+
+export function installTranscriptRecoveryRaceDom() {
+  const dom = new JSDOM('<!doctype html><html><body><div id="root"></div><div id="scroll"><div class="transcript__row" data-row-key="row-a"></div></div></body></html>', {
+    pretendToBeVisual: true,
+    url: "http://localhost/",
+  });
+  (globalThis as typeof globalThis & { IS_REACT_ACT_ENVIRONMENT: boolean }).IS_REACT_ACT_ENVIRONMENT = true;
+  globalThis.window = dom.window as unknown as Window & typeof globalThis;
+  globalThis.document = dom.window.document;
+  globalThis.HTMLElement = dom.window.HTMLElement;
+  globalThis.Element = dom.window.Element;
+  globalThis.Node = dom.window.Node;
+
+  let nextFrame = 1;
+  const frames = new Map<number, FrameRequestCallback>();
+  const requestFrame = (callback: FrameRequestCallback) => {
+    const id = nextFrame;
+    nextFrame += 1;
+    frames.set(id, callback);
+    return id;
+  };
+  const cancelFrame = (id: number) => void frames.delete(id);
+  globalThis.requestAnimationFrame = requestFrame;
+  globalThis.cancelAnimationFrame = cancelFrame;
+  dom.window.requestAnimationFrame = requestFrame;
+  dom.window.cancelAnimationFrame = cancelFrame;
+
+  const flushFrames = async () => {
+    const pending = [...frames.entries()];
+    frames.clear();
+    await act(async () => pending.forEach(([, callback]) => callback(performance.now())));
+  };
+  return { dom, flushFrames };
+}
