+++ desktop/frontend/src/lib/transcriptRows.ts
@@ -709,72 +709,6 @@ export function buildTranscriptRows(models: readonly TurnModel[], options: Build
   return rows;
 }
 
-// ── Live turn split ───────────────────────────────────────────────────────────
-// The streaming ("live") turn renders as the virtual list's in-flow Footer —
-// inside the scroller but outside the measured size tree — so the list only
-// ever owns static, bounded rows. The active turn is the one owning the live
-// assistant item; while running with no live item yet it is the last turn.
-// Everything from the active turn onward (except its user message, which
-// stays as the history tail) moves to the live region, preserving the visual
-// order for later steers/messages.
-
-export interface TranscriptLiveSplit {
-  historyRows: TranscriptRow[];
-  liveRows: TranscriptRow[];
-  /** True while a turn is active — the live region may render a status row even when liveRows is empty. */
-  liveActive: boolean;
-}
-
-export function resolveLiveTurnGrowthFloor(
-  previousRowCount: number,
-  nextRowCount: number,
-  previousHeight: number,
-  heldHeight: number | null,
-): number | null {
-  if (nextRowCount <= previousRowCount || previousHeight <= 0) return heldHeight;
-  return Math.max(previousHeight, heldHeight ?? 0);
-}
-
-function firstRowKeyForModel(model: TurnModel): string | undefined {
-  for (const segment of model.segments) {
-    if (segment.displayItems.length > 0) return `ph:${segment.key}`;
-    const outside = segment.outsideItems[0];
-    if (!outside) continue;
-    return outside.kind === "extension" ? `x:${outside.id}` : outside.kind === "notice" ? `n:${outside.id}` : `a:${outside.id}`;
-  }
-  return undefined;
-}
-
-export function splitTranscriptLiveRows(
-  models: readonly TurnModel[],
-  rows: readonly TranscriptRow[],
-  liveId: string | undefined,
-  running: boolean,
-): TranscriptLiveSplit {
-  let activeIndex = -1;
-  if (liveId) {
-    activeIndex = models.findIndex((model) => model.turnItems.some((item) => item.id === liveId));
-  }
-  if (activeIndex < 0 && running && models.length > 0) activeIndex = models.length - 1;
-  if (activeIndex < 0) return { historyRows: [...rows], liveRows: [], liveActive: false };
-  const active = models[activeIndex];
-  if (!active.user) {
-    // Prelude turn (no user message): split at the turn's first row. A turn
-    // with no rendered rows yet keeps the whole list as history.
-    const firstKey = firstRowKeyForModel(active);
-    const firstIndex = firstKey ? rows.findIndex((row) => row.key === firstKey) : -1;
-    if (!firstKey || firstIndex < 0) return { historyRows: [...rows], liveRows: [], liveActive: true };
-    return { historyRows: rows.slice(0, firstIndex), liveRows: rows.slice(firstIndex), liveActive: true };
-  }
-  const userIndex = rows.findIndex((row) => row.key === userRowKey(active.user!.id));
-  if (userIndex < 0) return { historyRows: [...rows], liveRows: [], liveActive: false };
-  return {
-    historyRows: rows.slice(0, userIndex + 1),
-    liveRows: rows.slice(userIndex + 1),
-    liveActive: true,
-  };
-}
-
 // ── Measurement / identity helpers ────────────────────────────────────────────
 
 /** Ballpark row heights; Virtuoso replaces them with real measurements on mount. */
